fix: funded by section
