+++ src/components/organisms/FundedBy.module.css
@@ -7,29 +7,22 @@
 .container {
   display: flex;
   flex-direction: column;
+  align-items: center;
 }
 
 .logo {
   display: flex;
   align-items: center;
   justify-content: center;
-}
-
-.poweredByContainer {
-  margin-top: var(--spacer);
+  max-width: 35rem;
 }
 
 @media screen and (min-width: 45rem) {
-  .wrapper {
-    display: grid;
-    grid-template-columns: 4fr 1fr;
-    gap: var(--spacer);
-  }
   .container {
-    flex-direction: row;
-    justify-content: space-between;
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
   }
-  .poweredByContainer {
-    margin-top: 0;
+  .logo {
+    max-width: none;
   }
 }
